requirements.txt
Flask==2.3.3
firebase-admin==6.2.0
python-dotenv==1.0.0
twilio==8.13.0
google-api-python-client>=2.120.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
google-auth>=2.23.4
pyOpenSSL==23.3.0
gunicorn==21.2.0  # just in linux for production
pytz>=2024.1

python-dotenv==1.0.0
